+++ source/tryGoal/component/tryGoalController.js
@@ -8,6 +8,7 @@ function($scope, agentMemory, $compile, $stateParams, config, GoalAPI, ConfigAPI
     $scope.otherOption = {value: '(other - not listed)'};
     $scope.yolandaUrl = ApiConfig.getConfig().url;
     $scope.tryGoal = agentMemory.tryGoal;
+    $scope.resultData;
 
     $scope.updateAlias = function() {
         sessionId = null;
@@ -91,10 +92,10 @@ function($scope, agentMemory, $compile, $stateParams, config, GoalAPI, ConfigAPI
     };
 
     $scope.$watch('apiOutput', function() {
-        
+
         if (agentMemory.tryGoal) {
             var displayText = JSON.stringify($rootScope.apiOutput, null, 1);
-            
+
             var lines = displayText.split('\n');
             lines.splice(0,1);
             if (lines.length > 0) {
@@ -106,7 +107,7 @@ function($scope, agentMemory, $compile, $stateParams, config, GoalAPI, ConfigAPI
             var logPanel = document.querySelector('#scroll-content');
             logPanel.scrollTop = logPanel.scrollHeight;
         }, 150);
-        
+
     }, true);
 
     $scope.toggleSplitScreen = function() {
@@ -180,7 +181,7 @@ function($scope, agentMemory, $compile, $stateParams, config, GoalAPI, ConfigAPI
         }
 
         $scope.postMessage(response);
-        
+
         if (response.question && response.question.allowUnknown) {
             focusElementById('mainAgent');
         }
@@ -248,26 +249,8 @@ function($scope, agentMemory, $compile, $stateParams, config, GoalAPI, ConfigAPI
             }
 
         } else if (response.result && angular.isArray(response.result) && response.result.length > 0) {
-            $scope.goalResults = [];
-            $scope.response = {};
-            response.result.forEach(function(element) {
-                var resultText = ($scope.goalInfo.goalText ? $scope.goalInfo.goalText : $scope.goalInfo.text);
-
-                var metaData = element.objectMetadata ? element.objectMetadata : '';
-
-                resultText = resultText.replace(/%O/g, element.object);
-                resultText = resultText.replace(/%R/g, element.relationship);
-                resultText = resultText.replace(/%S/g, element.subject);
-                resultText = resultText.replace(/%C/g, element.certainty);
-
-                if ($scope.config.showEvidence) {
-                    $scope.goalResults.push({text: resultText, cf: element.certainty, meta: metaData, factID: element.factID });
-                } else {
-                    $scope.goalResults.push({text: resultText, cf: element.certainty, meta: metaData});
-                }
-            });
+            $scope.resultData = response;
             $scope.display = 'result';
-
             if ($scope.tryGoal){
                 focusElementById('reset');
             } else {